+++ h2integrate/core/supported_models.py
@@ -1,13 +1,15 @@
 from h2integrate.resource.river import RiverResource
+from h2integrate.core.feedstocks import FeedstockCostModel, FeedstockPerformanceModel
 from h2integrate.transporters.pipe import PipePerformanceModel
 from h2integrate.transporters.cable import CablePerformanceModel
 from h2integrate.converters.steel.steel import SteelPerformanceModel, SteelCostAndFinancialModel
-from h2integrate.core.profast_financial import ProFastComp
 from h2integrate.converters.wind.wind_plant import WindPlantCostModel, WindPlantPerformanceModel
+from h2integrate.finances.profast_financial import ProFastComp
 from h2integrate.converters.hopp.hopp_wrapper import HOPPComponent
 from h2integrate.converters.solar.solar_pysam import PYSAMSolarPlantPerformanceModel
 from h2integrate.storage.hydrogen.eco_storage import H2Storage
 from h2integrate.converters.nitrogen.simple_ASU import SimpleASUCostModel, SimpleASUPerformanceModel
+from h2integrate.storage.simple_generic_storage import SimpleGenericStorage
 from h2integrate.storage.hydrogen.tank_baseclass import (
     HydrogenTankCostModel,
     HydrogenTankPerformanceModel,
@@ -18,6 +20,7 @@ from h2integrate.controllers.openloop_controllers import (
 )
 from h2integrate.converters.hydrogen.wombat_model import WOMBATElectrolyzerModel
 from h2integrate.converters.wind.wind_plant_pysam import PYSAMWindPlantPerformanceModel
+from h2integrate.storage.battery.atb_battery_cost import ATBBatteryCostModel
 from h2integrate.transporters.electricity_combiner import CombinerPerformanceModel
 from h2integrate.transporters.electricity_splitter import SplitterPerformanceModel
 from h2integrate.converters.ammonia.ammonia_synloop import (
@@ -35,6 +38,7 @@ from h2integrate.converters.hydrogen.pem_electrolyzer import (
 )
 from h2integrate.converters.solar.atb_res_com_pv_cost import ATBResComPVCostModel
 from h2integrate.converters.solar.atb_utility_pv_cost import ATBUtilityPVCostModel
+from h2integrate.resource.wind.nrel_developer_wtk_api import WTKNRELDeveloperAPIWindResource
 from h2integrate.converters.methanol.smr_methanol_plant import (
     SMRMethanolPlantCostModel,
     SMRMethanolPlantFinanceModel,
@@ -48,6 +52,10 @@ from h2integrate.converters.methanol.co2h_methanol_plant import (
     CO2HMethanolPlantCostModel,
     CO2HMethanolPlantFinanceModel,
     CO2HMethanolPlantPerformanceModel,
+)
+from h2integrate.converters.natural_gas.natural_gas_cc_ct import (
+    NaturalGasCostModel,
+    NaturalGasPerformanceModel,
 )
 from h2integrate.converters.hydrogen.singlitico_cost_model import SingliticoCostModel
 from h2integrate.converters.co2.marine.direct_ocean_capture import DOCCostModel, DOCPerformanceModel
@@ -81,6 +89,7 @@ from h2integrate.converters.hydrogen.geologic.stimulated_geoh2_plant import (
 supported_models = {
     # Resources
     "river_resource": RiverResource,
+    "wind_toolkit_v2_api": WTKNRELDeveloperAPIWindResource,
     # Converters
     "wind_plant_performance": WindPlantPerformanceModel,
     "wind_plant_cost": WindPlantCostModel,
@@ -125,20 +134,27 @@ supported_models = {
     "stimulated_geoh2_performance": StimulatedGeoH2PerformanceModel,
     "stimulated_geoh2_cost": StimulatedGeoH2CostModel,
     "stimulated_geoh2": StimulatedGeoH2FinanceModel,
+    "natural_gas_performance": NaturalGasPerformanceModel,
+    "natural_gas_cost": NaturalGasCostModel,
     # Transport
     "cable": CablePerformanceModel,
     "pipe": PipePerformanceModel,
     "combiner_performance": CombinerPerformanceModel,
     "splitter_performance": SplitterPerformanceModel,
+    # Storage
+    "h2_storage": H2Storage,
+    "hydrogen_tank_performance": HydrogenTankPerformanceModel,
+    "hydrogen_tank_cost": HydrogenTankCostModel,
+    "atb_battery_cost": ATBBatteryCostModel,
+    "simple_generic_storage": SimpleGenericStorage,
     # Control
     "pass_through_controller": PassThroughOpenLoopController,
     "demand_open_loop_controller": DemandOpenLoopController,
-    # Storage
-    "hydrogen_tank_performance": HydrogenTankPerformanceModel,
-    "hydrogen_tank_cost": HydrogenTankCostModel,
-    "h2_storage": H2Storage,
+    # Feedstock
+    "feedstock_performance": FeedstockPerformanceModel,
+    "feedstock_cost": FeedstockCostModel,
     # Finance
     "ProFastComp": ProFastComp,
 }
 
-electricity_producing_techs = ["wind", "solar", "pv", "river", "hopp"]
+electricity_producing_techs = ["wind", "solar", "pv", "river", "hopp", "natural_gas_plant"]